requirements.txt
langchain
langchain-openai
python-dotenv
flask
requests
reportlab
flask-limiter
apscheduler
waitress 